Move registration scenario password  algorithm to new registration component of common library
