+++ include/boe/registration.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string_view>
+
+short wd_to_pwd(std::string_view str);
