+++ src/Windows Code Release 3/Scen Ed/scenario.cpp
@@ -13,6 +13,7 @@
 #include "edsound.h"
 #include "boe/scenario.hpp"
 #include "boe/compatibility.hpp"
+#include "boe/registration.hpp"
 #include "scened_globals.hpp"
 
 
@@ -2769,42 +2770,4 @@ short enter_password()
 	cd_kill_dialog(823,0);
 
 	return i;
-}
-
-static const std::array<int, 6> c_pow{ 1,10,100,1000,9999,99999 };
-
-short wd_to_pwd(std::string_view str)
-{
-	char pwd[8] = "aaaaaa";
-	for (short i = 0; i < 6; ++i)
-	{
-		if (i >= str.size())
-		{
-			i = 6;
-		}
-		else
-		{
-			if ((str[i] >= 65) && (str[i] <= 90))
-			{
-				pwd[i] = str[i] + 32;
-			}
-			else if ((str[i] >= 48) && (str[i] <= 57))
-			{
-				pwd[i] = str[i] + 49;
-			}
-			else if ((str[i] >= 97) && (str[i] <= 122))
-			{
-				pwd[i] = str[i];
-			}
-		}
-	}
-
-	int val = 0;
-	for (short i = 0; i < 6; ++i)
-	{
-		val += c_pow[i] * (int)(pwd[i] - 97);
-	}
-	val %= 30000;
-
-	return (short)val;
 }
