+++ src/Windows Code Release 3/Scen Ed/scenario.h
@@ -60,6 +60,4 @@ void set_starting_loc_filter (short item_hit);
 void set_starting_loc();
 void edit_boats_event_filter (short item_hit);
 void edit_boats();
-
 void user_password_filter (short item_hit);
-short wd_to_pwd(std::string_view str);
